refactor: import component from next-cloudinary in reusable component and extract header into it's own component
