+++ examples/nextjs-cldimage/app/layout.jsx
@@ -1,4 +1,12 @@
+import Header from '../components/Header';
 import '../styles/globals.css';
+import { Inter } from 'next/font/google';
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  weight: ['400', '700'],
+});
 
 export const metadata = {
   title: 'Next Cloudinary CldImage Component',
@@ -9,7 +17,10 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang='en'>
-      <body>{children}</body>
+      <body className={inter.className}>
+        <Header />
+        {children}
+      </body>
     </html>
   );
 }
