+++ examples/nextjs-cldimage/pages/_app.jsx
@@ -1,7 +1,20 @@
+import Header from '../components/Header';
 import '../styles/globals.css';
+import { Inter } from 'next/font/google';
+
+const inter = Inter({
+  subsets: ['latin'],
+  display: 'swap',
+  weight: ['400', '700'],
+});
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+  return (
+    <main className={inter.className}>
+      <Header />
+      <Component {...pageProps} />
+    </main>
+  );
 }
 
 export default MyApp;
